+++ bugsnag-plugin-android-ndk/src/main/assets/include/bugsnag.h
@@ -16,29 +16,38 @@ typedef bool (*bsg_on_error)(void *);
  * @param env  The JNI environment to use when using convenience methods
  */
 void bugsnag_start(JNIEnv *env);
+
 /**
  * Sends an error report to Bugsnag
  * @param name     The name of the error
  * @param message  The error message
  * @param severity The severity of the error
  */
-void bugsnag_notify(char *name, char *message, bugsnag_severity severity);
-void bugsnag_notify_env(JNIEnv *env, char *name, char *message,
+void bugsnag_notify(const char *name, const char *message,
+                    bugsnag_severity severity);
+
+void bugsnag_notify_env(JNIEnv *env, const char *name, const char *message,
                         bugsnag_severity severity);
+
 /**
  * Set the current user
  * @param id    The identifier of the user
  * @param email The user's email
  * @param name  The user's name
  */
-void bugsnag_set_user(char *id, char *email, char *name);
-void bugsnag_set_user_env(JNIEnv *env, char *id, char *email, char *name);
+void bugsnag_set_user(const char *id, const char *email, const char *name);
+
+void bugsnag_set_user_env(JNIEnv *env, const char *id, const char *email,
+                          const char *name);
+
 /**
  * Leave a breadcrumb, indicating an event of significance which will be logged
  * in subsequent error reports
  */
-void bugsnag_leave_breadcrumb(char *message, bugsnag_breadcrumb_type type);
-void bugsnag_leave_breadcrumb_env(JNIEnv *env, char *message,
+void bugsnag_leave_breadcrumb(const char *message,
+                              bugsnag_breadcrumb_type type);
+
+void bugsnag_leave_breadcrumb_env(JNIEnv *env, const char *message,
                                   bugsnag_breadcrumb_type type);
 
 /**